requirements.txt
absl-py==2.3.0
accelerate==1.8.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anthropic==0.55.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appdirs==1.4.4
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
arxiv==2.2.0
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
azure-ai-inference==1.0.0b9
azure-core==1.34.0
azure-core-tracing-opentelemetry==1.0.0b12
azure-cosmos==4.9.0
azure-identity==1.23.0
babel==2.17.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.2.1
cohere==5.15.0
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.2
cryptography==45.0.4
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.6.0
debugpy==1.8.14
decorator==5.2.1
deepl==1.22.0
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.8
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
docx2txt==0.9
durationpy==0.10
effdet==0.4.1
elastic-transport==8.17.1
elasticsearch==8.18.1
emoji==2.14.1
et_xmlfile==2.0.0
executing==2.2.0
faiss-cpu==1.11.0
fastapi==0.115.14
fastavro==1.11.1
fastjsonschema==2.21.1
feedparser==6.0.11
filelock==3.18.0
filetype==1.2.0
FlashRank==0.2.10
flatbuffers==25.2.10
fonttools==4.58.4
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.3.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.18
google-api-core==1.34.1
google-auth==2.40.3
google-cloud-vision==3.10.2
google_search_results==2.4.2
googleapis-common-protos==1.70.0
grandalf==0.8
greenlet==3.2.3
grpc-gateway-protoc-gen-openapiv2==0.1.0
grpcio==1.67.1
grpcio-status==1.48.2
h11==0.16.0
hf-xet==1.1.5
html5lib==1.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isodate==0.7.2
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jpype1==1.5.2
json5==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwipiepy==0.21.0
kiwipiepy_model==0.21.0
kiwisolver==1.4.8
konlpy==0.6.0
kubernetes==33.1.0
langchain==0.3.26
langchain-anthropic==0.3.16
langchain-azure-ai==0.1.4
langchain-chroma==0.2.3
langchain-cohere==0.4.4
langchain-community==0.3.26
langchain-core==0.3.66
langchain-elasticsearch==0.3.2
langchain-experimental==0.3.4
langchain-google-genai==2.1.5
langchain-huggingface==0.3.0
langchain-milvus==0.2.1
langchain-ollama==0.3.3
langchain-openai==0.3.27
langchain-teddynote==0.3.45
langchain-text-splitters==0.3.8
langchainhub==0.1.21
langdetect==1.0.9
langgraph==0.5.0
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.5.1
langgraph-sdk==0.1.72
langsmith==0.4.4
lark==1.2.2
llama-cloud==0.1.19
llama-cloud-services==0.6.21
llama-index-core==0.11.23
llama-index-readers-file==0.2.2
llama-parse==0.6.21
lxml==5.4.0
lz4==4.4.4
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
milvus-lite==2.5.0
mistune==3.1.3
mmh3==4.1.0
mpmath==1.3.0
msal==1.32.3
msal-extensions==1.3.1
multidict==6.6.2
multiprocess==0.70.16
mypy==1.16.1
mypy_extensions==1.1.0
narwhals==1.44.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
notebook==7.4.3
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.3.1
olefile==0.47
ollama==0.5.1
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.22.0
open_clip_torch==2.32.0
openai==1.93.0
opencv-python==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.15.0
opentelemetry-instrumentation==0.55b1
opentelemetry-instrumentation-asgi==0.55b1
opentelemetry-instrumentation-fastapi==0.55b1
opentelemetry-proto==1.15.0
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
opentelemetry-util-http==0.55b1
orjson==3.10.18
ormsgpack==1.10.0
overrides==7.7.0
packaging==24.2
pandas==2.3.0
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.5
pexpect==4.9.0
pi_heif==0.22.0
pikepdf==9.9.0
pillow==10.4.0
pinecone==5.4.2
pinecone-client==3.2.2
pinecone-plugin-inference==3.1.0
pinecone-plugin-interface==0.0.7
pinecone-text==0.10.0
platformdirs==4.3.8
posthog==6.0.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==3.20.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pycocotools==2.0.10
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.9.0
pymilvus==2.5.11
PyMuPDF==1.26.1
pymupdf4llm==0.0.25
pypandoc==1.15
pyparsing==3.2.3
pypdf==4.3.1
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-iso639==2025.2.18
python-json-logger==3.3.0
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
ragas==0.2.15
rank-bm25==0.2.2
RapidFuzz==3.13.0
rapidocr-onnxruntime==1.4.4
redis==5.3.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rouge_score==0.1.2
rpds-py==0.25.1
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
seaborn==0.13.2
Send2Trash==1.8.3
sgmllib3k==1.0.0
shapely==2.1.1
shellingham==1.5.4
simsimd==6.4.9
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.46.2
streamlit==1.46.1
striprtf==0.0.26
sympy==1.14.0
tavily-python==0.7.8
tenacity==8.5.0
terminado==0.18.1
threadpoolctl==3.6.0
tiktoken==0.9.0
timm==1.0.16
tinycss2==1.4.0
tokenizers==0.21.2
toml==0.10.2
torch==2.2.0
torchvision==0.17.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.0
typer==0.16.0
types-python-dateutil==2.9.0.20250516
types-PyYAML==6.0.12.20250516
types-requests==2.32.4.20250611
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ujson==5.10.0
unstructured==0.18.1
unstructured-client==0.37.2
unstructured-inference==1.0.5
unstructured.pytesseract==0.3.15
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
widgetsnbextension==4.0.14
wikipedia==1.4.0
wrapt==1.17.2
xlrd==2.0.2
xlsxwriter==3.2.5
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
